Working partial resolution to issue # 29 (Gear Lockers).
DOES - Players should be able to store gear in lockers.
DOES NOT - Lockers should be an interactable that, when stepped on, activates a certain locker.
DOES NOT - Each locker should have an identifier that links it to other lockers of the same type (lockers will be regional, most likely).
DOES - There should be a way to prevent certain items from entering the locker (player.canLocker(item)). For example, no corpses or gold.
DOES - This should support all of the drag and drop functionality, with a new set of commands (XtK/KtX where K is short for locKer since L is already taken).
DOES - There needs to be a new option for Locker size (normal, small, mini).
DOES - Lockers should store 20 items.